Fix the memory corruption by integer overflow.
To get all items from a list, the following list index range can be used.
- 0..2147483647

The process_lop_get function estimates how many elements can be gotten
and allocates the memory for saving the addresses of elements to be retrived.
In estimating the element count, we use the list index range like below
- estimated_count = 2147483647(to_index) - 0(from_index) + 1

The process_lop_get fucntion used an integer varible to save
the result of above computation. And it generates an integer overflow.
This overflowed integer value makes smaller memory to be allocated than needed,
and the function overwrites across the boundary of the allocate memory.
This makes memory corruption..

To avoid the integer overflow, we adjusted the list indexes not to be large values.
Refer to the modified code.