prepend "mc_" prefix to hash() function name.
arcus-memcached has isasl functionality when build it with --enable-isasl option.
The build process makes a compile error which is an implicit declaration of function 'hash'.
To avoid that, we changed the function name to 'mc_hash()" in isasl code.

Refere to the commit f854ad8702fd44b7652c4f0262d7e173b12d9f07.