EDUCE_EVICTION_SPIKE: reduce too many eviciton occurrences.
In low space shortage level, that is 1 ~ 9,
only worker threads can evict cache items to obtain free space.
Therefore, eviction ratio will be proportional to the amount
of small meory requests.
In high space shortage level, that is 10 ~ 100,
a background daemon thread as well as worker threads
evict cache items to obtain enough free space.
In this case, too many evictions can occur.
But, it's unavoidable to prevent an "out of memory" error.