JHPARK_NEW_SMGET_INTERFACE: fix segfault.
The segfault was caused by applying minus operator on unsigned integer.
The trim_elems array of smget result structure is used as a backward array.
And, we access each trim element with the index come out
by applying minus operator on trim_count of uint32_t type.
In that case, the computed index was not the intended index.